--- job.css
@@ -1,60 +1,0 @@
-.job-form{
-	    width: 380px;
-	    background: #cde0e2;
-	    height: 460px;
-	    padding: 50px 70px;
-	    border-radius: 10px;
-	    position: absolute;
-	    left: 50%;
-	    top: 58%;
-	    transform: translate(-50%,-50%);
-	  }
-	  .job-form h1{
-	    text-align: center;
-	    margin-bottom: 20px;
-	    color: #201e3b;
-	  }
-	  .description input{
-		height: 150px;
-		width: 300px;
-		border: 2px solid #272253;
-		resize: none;
-		margin: 20px 0;
-	  }
-	  .txtbr{
-	    border-bottom: 2px solid #272253;
-	    position: relative;
-	    margin: 20px 0;
-	  }
-	  .txtbr input{
-	    font-size: 15px;
-	    color: #333;
-	    border: none;
-	    width: 100%;
-	    outline: none;
-	    background: none;
-	    padding: 0 5px;
-	    height: 40px;
-	  }
-	  .jobbtn{
-	    display: block;
-	    width: 100%;
-	    height: 50px;
-	    border: none;
-	    background: linear-gradient(120deg,#34dbb7,#8e44ad,#3498db);
-	    background-size: 200%;
-	    color: #fff;
-	    outline: none;
-	    cursor: pointer;
-	    transition: .5s;
-	  }
-	  .jobbtn:hover{
-	    background-position: right;
-	  }
-	  .bottom-text{
-	    margin-top: 10px;
-	    text-align: center;
-	    font-size: 13px;
-	    color: black;
-	  }
-
